embed(A1): event serde, content_index u32, Arc EventSink + subscribe_sink
Layer A1 foundation for non-Rust embedding (additive; Rust-native APIs unchanged):

- AgentEvent derives Serialize/Deserialize (+ round-trip test) -> stable wire/
  replay format. Surfaced and fixed a pre-existing genai::ModelName serde bug:
  it serialized as {"Shared":..} but only deserialized from a bare string, so no
  message-carrying event round-tripped. Serialize now emits a bare string matching
  Deserialize; Deserialize accepts owned strings (from_reader/escaped). All 9
  yakbak provider suites stay green (wire path unaffected).
- content_index: usize -> u32 on all nine AssistantMessageEvent streaming-delta
  fields (text/thinking/tool-call x start/delta/end); usize has no UniFFI mapping.
  25 construction casts; Vec-index locals stay usize; behavior identical.
- EventSink (Arc/&self, async) + Agent::subscribe_sink, reusing the awaited-
  listener machinery so emit is sequentially awaited (backpressure) on worker
  threads. Test proves the sink observes the full awaited event sequence.

docs/embedding.md: §3.4 corrected to nine content_index fields.

Gates: workspace check clean; rust-genai-agent --all-features 198 passed;
genai-agentprism lib 284, codex 316 + codex_stream 10; all 9 yakbak suites green.

+++ agent/src/proxy/accumulator.rs
@@ -197,7 +197,7 @@ impl ProxyAccumulator {
         self.partial.content.push(AssistantContent::text(""));
         self.blocks.push(BlockState::Text { open: true });
         vec![AssistantMessageEvent::TextStart {
-            content_index: index,
+            content_index: index as u32,
             partial: self.partial.clone(),
         }]
     }
@@ -217,7 +217,7 @@ impl ProxyAccumulator {
         };
         text.push_str(&delta);
         vec![AssistantMessageEvent::TextDelta {
-            content_index: index,
+            content_index: index as u32,
             delta,
             partial: self.partial.clone(),
         }]
@@ -247,7 +247,7 @@ impl ProxyAccumulator {
             *open = false;
         }
         vec![AssistantMessageEvent::TextEnd {
-            content_index: index,
+            content_index: index as u32,
             content,
             partial: self.partial.clone(),
         }]
@@ -261,7 +261,7 @@ impl ProxyAccumulator {
         self.partial.content.push(AssistantContent::thinking(""));
         self.blocks.push(BlockState::Thinking { open: true });
         vec![AssistantMessageEvent::ThinkingStart {
-            content_index: index,
+            content_index: index as u32,
             partial: self.partial.clone(),
         }]
     }
@@ -286,7 +286,7 @@ impl ProxyAccumulator {
         };
         thinking.push_str(&delta);
         vec![AssistantMessageEvent::ThinkingDelta {
-            content_index: index,
+            content_index: index as u32,
             delta,
             partial: self.partial.clone(),
         }]
@@ -319,7 +319,7 @@ impl ProxyAccumulator {
             *open = false;
         }
         vec![AssistantMessageEvent::ThinkingEnd {
-            content_index: index,
+            content_index: index as u32,
             thinking,
             partial: self.partial.clone(),
         }]
@@ -348,7 +348,7 @@ impl ProxyAccumulator {
             delta_count: 0,
         });
         vec![AssistantMessageEvent::ToolCallStart {
-            content_index: index,
+            content_index: index as u32,
             partial: self.partial.clone(),
         }]
     }
@@ -431,7 +431,7 @@ impl ProxyAccumulator {
         };
         tool_call.arguments = parse_streaming_json(&raw_arguments);
         vec![AssistantMessageEvent::ToolCallDelta {
-            content_index: index,
+            content_index: index as u32,
             delta,
             partial: self.partial.clone(),
         }]
@@ -461,7 +461,7 @@ impl ProxyAccumulator {
             *open = false;
         }
         vec![AssistantMessageEvent::ToolCallEnd {
-            content_index: index,
+            content_index: index as u32,
             tool_call,
             partial: self.partial.clone(),
         }]
